refactor: Create generic RadialIntensityIndicator component
Extracts the `BrightnessIndicator` from the Torch screen into a more generic and reusable `RadialIntensityIndicator` component.

- Creates `RadialIntensityIndicator.kt` in the `ui/components` package.
- Renames the composable and its `brightness` parameter to the more general `intensity`.
- Deletes the old `BrightnessIndicator.kt` file.
- Updates `TorchScreen` and `MagnetometerScreen` to use the new `RadialIntensityIndicator` component, adapting it for their specific use cases (torch brightness and magnetic field strength).
