+++ app/src/main/java/com/shresht7/pockettools/ui/components/RadialIntensityIndicator.kt
@@ -1,4 +1,4 @@
-package com.shresht7.pockettools.ui.screens.torch
+package com.shresht7.pockettools.ui.components
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.fillMaxSize
@@ -12,8 +12,8 @@ import kotlin.math.cos
 import kotlin.math.sin
 
 @Composable
-fun BrightnessIndicator(
-    brightness: Float,
+fun RadialIntensityIndicator(
+    intensity: Float,
     innerRadiusFactor: Float,
     outerRadiusFactor: Float,
     center: Offset? = null,
@@ -29,9 +29,9 @@ fun BrightnessIndicator(
 
         repeat(steps) { i ->
             val threshold = (i + 1) / steps.toFloat()
-            if (brightness >= threshold - 0.001f) {
+            if (intensity >= threshold - 0.001f) {
                 val radius = buttonRadius + maxRadius * ((i + 1) / steps.toFloat())
-                val alpha = (brightness - (threshold - 0.2f).coerceIn(0f, 1f))
+                val alpha = (intensity - (threshold - 0.2f).coerceIn(0f, 1f))
 
                 for (j in 0 until dotsPerCircle) {
                     val angle = (2 * Math.PI * j / dotsPerCircle).toFloat()
@@ -50,9 +50,9 @@ fun BrightnessIndicator(
 
 @Preview(showBackground = true)
 @Composable
-fun BrightnessIndicatorPreview() {
-    BrightnessIndicator(
-        brightness = 0.5f,
+fun RadialIntensityIndicatorPreview() {
+    RadialIntensityIndicator(
+        intensity = 0.5f,
         innerRadiusFactor = 0.5f,
         outerRadiusFactor = 0.9f,
         steps = 5,
